requirements.txt
streamlit==1.21.0
streamlit-folium==0.12.0
pandas==1.5.3
numpy==1.21.2
folium==0.14.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
haversine==2.8.0
Pillow==9.0.1
plotly==5.14.1